Add commit
Add --op=commit to create commits in repo. Parameters are
--num-commits (total commits to create), --adds-per-commit
(number of add operations for each commit), and --files-per-add
(number of files in a single add operation). This flexibility
is to allow testing of bulk operations.

Refactor worktree. Now, a Repo has a worktree, and defers file
operations to it.

Redo options a bit, although this is still in flux, looking for
the right balance of explicit vs easy to use.